requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.3
SQLAlchemy==1.4.41
Faker==15.3.2
PyJWT==2.6.0
